Move the code for handling areas, markers and polylines to separate files
